Polished and moved ZipEntryTreeBuilder to new package
