+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/ui/inject/Injector.kt
@@ -1,6 +1,6 @@
 package org.wycliffeassociates.otter.jvm.app.ui.inject
 
-import org.wycliffeassociates.otter.jvm.ZipEntryTreeBuilder
+import org.wycliffeassociates.otter.jvm.domain.resourceontainer.project.ZipEntryTreeBuilder
 import org.wycliffeassociates.otter.jvm.device.audio.injection.DaggerAudioComponent
 import org.wycliffeassociates.otter.jvm.device.audioplugin.injection.DaggerAudioPluginComponent
 import org.wycliffeassociates.otter.jvm.persistence.injection.DaggerPersistenceComponent
